added window stats to cc_traveltime_misfit function, moved cc_error and utility functions from cc traveltime misfit into the utils directory to clean things up
