+++ pyadjoint/adjoint_source_types/cc_traveltime_misfit.py
@@ -9,12 +9,11 @@ Cross correlation traveltime misfit and associated adjoint source.
 :license:
     BSD 3-Clause ("BSD New" or "BSD Simplified")
 """
-import warnings
 import numpy as np
-from obspy.signal.cross_correlation import xcorr_pick_correction
 from scipy.integrate import simps
 
-from pyadjoint.utils import window_taper,  generic_adjoint_source_plot
+from pyadjoint.utils import (window_taper, generic_adjoint_source_plot,
+                             xcorr_shift, cc_error)
 
 
 VERBOSE_NAME = "Cross Correlation Traveltime Misfit"
@@ -86,7 +85,7 @@ def calculate_adjoint_source(observed, synthetic, config, window,
     :param observed: observed waveform to calculate adjoint source
     :type synthetic:  obspy.core.trace.Trace
     :param synthetic: synthetic waveform to calculate adjoint source
-    :type config: pyadjoint.config.ConfigWaveform
+    :type config: pyadjoint.config.ConfigCCTraveltime
     :param config: Config class with parameters to control processing
     :type window: list of tuples
     :param window: [(left, right),...] representing left and right window
@@ -101,9 +100,15 @@ def calculate_adjoint_source(observed, synthetic, config, window,
     assert(config.__class__.__name__ == "ConfigCCTraveltime"), \
         "Incorrect configuration class passed to CCTraveltime misfit"
 
+    # Allow for measurement types related to `dt` (p) and `dlna` (q)
     ret_val_p = {}
     ret_val_q = {}
 
+    # List of windows and some measurement values for each
+    win_stats_p = []
+    win_stats_q = []
+
+    # Initiate constants and empty return values to fill
     nlen_data = len(synthetic.data)
     deltat = synthetic.stats.delta
 
@@ -135,21 +140,29 @@ def calculate_adjoint_source(observed, synthetic, config, window,
         window_taper(s, taper_percentage=config.taper_percentage,
                      taper_type=config.taper_type)
 
-        i_shift = _xcorr_shift(d, s)
+        i_shift = xcorr_shift(d, s)
         t_shift = i_shift * deltat
 
         cc_dlna = 0.5 * np.log(sum(d[0:nlen] * d[0:nlen]) /
                                sum(s[0:nlen] * s[0:nlen]))
 
-        sigma_dt, sigma_dlna = cc_error(
-            d1=d, d2=s, deltat=deltat, cc_shift=i_shift, cc_dlna=cc_dlna,
-            dt_sigma_min=config.dt_sigma_min,
-            dlna_sigma_min=config.dlna_sigma_min
-        )
+        # Determine if a calculated error value should be used for normalization
+        if config.use_cc_error:
+            sigma_dt, sigma_dlna = cc_error(
+                d1=d, d2=s, deltat=deltat, cc_shift=i_shift, cc_dlna=cc_dlna,
+                dt_sigma_min=config.dt_sigma_min,
+                dlna_sigma_min=config.dlna_sigma_min
+            )
+        else:
+            sigma_dt, sigma_dlna = 1., 1.
 
-        misfit_sum_p += 0.5 * (t_shift / sigma_dt) ** 2
-        misfit_sum_q += 0.5 * (cc_dlna / sigma_dlna) ** 2
+        # Calculate the misfit for both time shift and amplitude anomaly
+        misfit_p = 0.5 * (t_shift / sigma_dt) ** 2
+        misfit_q = 0.5 * (cc_dlna / sigma_dlna) ** 2
+        misfit_sum_p += misfit_p
+        misfit_sum_q += misfit_q
 
+        # Calculate adjoint sources for both time shift and amplitude anomaly
         dsdt = np.gradient(s, deltat)
         nnorm = simps(y=dsdt * dsdt, dx=deltat)
         fp[left_sample:right_sample] = dsdt[:] * t_shift / nnorm / sigma_dt ** 2
@@ -158,126 +171,41 @@ def calculate_adjoint_source(observed, synthetic, config, window,
         fq[left_sample:right_sample] = \
             -1.0 * s[:] * cc_dlna / mnorm / sigma_dlna ** 2
 
+        # Store some information for each window
+        win_stats_q.append(
+            {"left": left_window_border, "right": right_window_border,
+             "measurement_type": config.measure_type,
+             "dlna": cc_dlna,  "misfit_dlna": misfit_q,
+             "sigma_dlna": sigma_dlna,
+             }
+        )
+        win_stats_q.append(
+            {"left": left_window_border, "right": right_window_border,
+             "measurement_type": config.measure_type,
+             "tshift": t_shift,  "misfit_dt": misfit_p, "sigma_dt": sigma_dt,
+             }
+        )
+
+    # Keep track of both misfit values
     ret_val_p["misfit"] = misfit_sum_p
     ret_val_q["misfit"] = misfit_sum_q
 
     if adjoint_src is True:
         ret_val_p["adjoint_source"] = fp[::-1]
         ret_val_q["adjoint_source"] = fq[::-1]
+    if window_stats is True:
+        ret_val_p["window_stats"] = win_stats_p
+        ret_val_q["window_stats"] = win_stats_q
 
     if config.measure_type == "dt":
-        if plot:
-            generic_adjoint_source_plot(observed, synthetic,
-                                        ret_val_p["adjoint_source"],
-                                        ret_val_p["misfit"],
-                                        window, VERBOSE_NAME)
+        ret_val = ret_val_p
+    elif config.measure_type == "am":
+        ret_val = ret_val_q
 
-        return ret_val_p
+    if plot:
+        generic_adjoint_source_plot(
+            observed, synthetic, ret_val["adjoint_source"], ret_val["misfit"],
+            window, VERBOSE_NAME
+        )
 
-    if config.measure_type == "am":
-        if plot:
-            generic_adjoint_source_plot(observed, synthetic,
-                                        ret_val_q["adjoint_source"],
-                                        ret_val_q["misfit"],
-                                        window, VERBOSE_NAME)
-
-        return ret_val_q
-
-
-def subsample_xcorr_shift(d, s):
-    """
-    Calculate the correlation time shift around the maximum amplitude of the
-    synthetic trace `s` with subsample accuracy.
-
-    :type d: obspy.core.trace.Trace
-    :param d: observed waveform to calculate adjoint source
-    :type s:  obspy.core.trace.Trace
-    :param s: synthetic waveform to calculate adjoint source
-    """
-    # Estimate shift and use it as a guideline for the subsample accuracy shift.
-    time_shift = _xcorr_shift(d.data, s.data) * d.stats.delta
-
-    # Align on the maximum amplitude of the synthetics.
-    pick_time = s.stats.starttime + s.data.argmax() * s.stats.delta
-
-    # Will raise a warning if the trace ids don't match which we don't care
-    # about here.
-    with warnings.catch_warnings():
-        warnings.simplefilter("ignore")
-        return xcorr_pick_correction(
-            pick_time, s, pick_time, d, 20.0 * time_shift,
-            20.0 * time_shift, 10.0 * time_shift)[0]
-
-
-def _xcorr_shift(d, s):
-    """
-    Determine the required time shift for peak cross-correlation of two arrays
-
-    :type d: np.array
-    :param d: observed time series array
-    :type s:  np.array
-    :param s: synthetic time series array
-    """
-    cc = np.correlate(d, s, mode="full")
-    time_shift = cc.argmax() - len(d) + 1
-    return time_shift
-
-
-def cc_error(d1, d2, deltat, cc_shift, cc_dlna, dt_sigma_min, dlna_sigma_min):
-    """
-    Estimate error for `dt` and `dlna` with uncorrelation assumption. Used for
-    normalization of the traveltime measurement
-
-    :type d1: np.array
-    :param d1: time series array to calculate error for
-    :type d2: np.array
-    :param d2: time series array to calculate error for
-    :type cc_shift: int
-    :param cc_shift: total amount of cross correlation time shift
-    :type cc_dlna: float
-    :param cc_dlna: amplitude anomaly calculated for cross-correlation
-        measurement
-    :type dt_sigma_min: float
-    :param dt_sigma_min: minimum travel time error allowed
-    :type dlna_sigma_min: float
-    :param dlna_sigma_min: minimum amplitude error allowed
-    """
-    nlen_t = len(d1)
-
-    d2_cc_dt = np.zeros(nlen_t)
-    d2_cc_dtdlna = np.zeros(nlen_t)
-
-    for index in range(0, nlen_t):
-        index_shift = index - cc_shift
-
-        if 0 <= index_shift < nlen_t:
-            # corrected by c.c. shift
-            d2_cc_dt[index] = d2[index_shift]
-
-            # corrected by c.c. shift and amplitude
-            d2_cc_dtdlna[index] = np.exp(cc_dlna) * d2[index_shift]
-
-    # time derivative of d2_cc (velocity)
-    d2_cc_vel = np.gradient(d2_cc_dtdlna, deltat)
-
-    # the estimated error for dt and dlna with uncorrelation assumption
-    sigma_dt_top = np.sum((d1 - d2_cc_dtdlna)**2)
-    sigma_dt_bot = np.sum(d2_cc_vel**2)
-
-    sigma_dlna_top = sigma_dt_top
-    sigma_dlna_bot = np.sum(d2_cc_dt**2)
-
-    sigma_dt = np.sqrt(sigma_dt_top / sigma_dt_bot)
-    sigma_dlna = np.sqrt(sigma_dlna_top / sigma_dlna_bot)
-
-    if sigma_dt < dt_sigma_min:
-        sigma_dt = dt_sigma_min
-
-    if sigma_dlna < dlna_sigma_min:
-        sigma_dlna = dlna_sigma_min
-
-    return sigma_dt, sigma_dlna
-
-
-
-
+    return ret_val
